// frontend/src/apisv2beta1/run/models/InputOutputsIOArtifact.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2beta1/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { V2beta1IOType } from './V2beta1IOType';
import {
  V2beta1IOTypeFromJSON,
  V2beta1IOTypeFromJSONTyped,
  V2beta1IOTypeToJSON,
  V2beta1IOTypeToJSONTyped,
} from './V2beta1IOType';
import type { V2beta1Artifact } from './V2beta1Artifact';
import {
  V2beta1ArtifactFromJSON,
  V2beta1ArtifactFromJSONTyped,
  V2beta1ArtifactToJSON,
  V2beta1ArtifactToJSONTyped,
} from './V2beta1Artifact';
import type { V2beta1IOProducer } from './V2beta1IOProducer';
import {
  V2beta1IOProducerFromJSON,
  V2beta1IOProducerFromJSONTyped,
  V2beta1IOProducerToJSON,
  V2beta1IOProducerToJSONTyped,
} from './V2beta1IOProducer';

/**
 *
 * @export
 * @interface InputOutputsIOArtifact
 */
export interface InputOutputsIOArtifact {
  /**
   *
   * @type {Array<V2beta1Artifact>}
   * @memberof InputOutputsIOArtifact
   */
  artifacts?: Array<V2beta1Artifact>;
  /**
   *
   * @type {V2beta1IOType}
   * @memberof InputOutputsIOArtifact
   */
  type?: V2beta1IOType;
  /**
   *
   * @type {string}
   * @memberof InputOutputsIOArtifact
   */
  artifact_key?: string;
  /**
   *
   * @type {V2beta1IOProducer}
   * @memberof InputOutputsIOArtifact
   */
  producer?: V2beta1IOProducer;
}

/**
 * Check if a given object implements the InputOutputsIOArtifact interface.
 */
export function instanceOfInputOutputsIOArtifact(value: object): value is InputOutputsIOArtifact {
  return true;
}

export function InputOutputsIOArtifactFromJSON(json: any): InputOutputsIOArtifact {
  return InputOutputsIOArtifactFromJSONTyped(json, false);
}

export function InputOutputsIOArtifactFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): InputOutputsIOArtifact {
  if (json == null) {
    return json;
  }
  return {
    artifacts:
      json['artifacts'] == null
        ? undefined
        : (json['artifacts'] as Array<any>).map(V2beta1ArtifactFromJSON),
    type: json['type'] == null ? undefined : V2beta1IOTypeFromJSON(json['type']),
    artifact_key: json['artifact_key'] == null ? undefined : json['artifact_key'],
    producer: json['producer'] == null ? undefined : V2beta1IOProducerFromJSON(json['producer']),
  };
}

export function InputOutputsIOArtifactToJSON(json: any): InputOutputsIOArtifact {
  return InputOutputsIOArtifactToJSONTyped(json, false);
}

export function InputOutputsIOArtifactToJSONTyped(
  value?: InputOutputsIOArtifact | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    artifacts:
      value['artifacts'] == null
        ? undefined
        : (value['artifacts'] as Array<any>).map(V2beta1ArtifactToJSON),
    type: V2beta1IOTypeToJSON(value['type']),
    artifact_key: value['artifact_key'],
    producer: V2beta1IOProducerToJSON(value['producer']),
  };
}